resolver: inline small quic module
